Fork update. (#1)
* SF.12: Prefer the quoted form of `#include` for files relative to the including file

The current guidance on SF.12 can be over-applied and devolves into "always use <>" because all compilers support adding include directories to the <> search. In this case, even the current directory may be added and so it is always possible to use <> for every header. Applying the guidance then devolves into an undesirable state where <> is always used and include"" is never used.

Instead, the proposed guidance leverages and encourages the distinction between <> and "" to create an easy-to-understand rule that the original guidance hints at and that most developers already follow and understand: "" is for local headers and <> is for library and external headers.

* Update wording and examples for SF.12: Prefer the quoted form of `#include` for files relative 

Updated wording, adjusted example, and provided a more verbose example of what can go wrong if using the wrong form.

* update code markdown for #include

* updated wording around use of <> for projects

make it more clear that using <> for external projects is just a typical example (and not the only use case for <> outside of standard headers)

* keep wording more consistent

.... if a bit more monotonous, but that's ok.

* Update CppCoreGuidelines.md

* Update CppCoreGuidelines.md

explicitly note that "file" refers to as it exists in the location of it being authored/modified.
this is to address any confusion about #including "bar.h" from foo.h based on whether bar.h is relative to foo.h at the time and location foo.h is being modified (e.g. under mylib/foo.h) versus at the time/location from which foo.h is installed to the system and included by the user (e.g. from /usr/include).

* relative path should be from within the same component

... files that exist at a relative path to the file containing the `#include` statement (from within the same component or project)

Co-authored-by: apenn-msft <62863214+apenn-msft@users.noreply.github.com>
Co-authored-by: Gabriel Dos Reis <gdr-at-ms@users.noreply.github.com>